fixing test failures and cleanups
also moved common test cluster start/stop etc to new class ClusterUtils that also
creates a separate directory for cluster instead of using the main snappy product directory

+++ cluster/src/dunit/scala/org/apache/spark/memory/SnappyUnifiedMemoryManagerDUnitTest.scala
@@ -109,9 +109,11 @@ class SnappyUnifiedMemoryManagerDUnitTest(s: String) extends ClusterManagerTestB
   }
 
   private def cleanTestResources(): Unit = {
-    val snc = SnappyContext(sc).newSession()
-    snc.dropTable(col_table, ifExists = true)
-    snc.dropTable(rr_table, ifExists = true)
+    val snc = SnappyContext(sc)
+    if (snc ne null) {
+      snc.dropTable(col_table, ifExists = true)
+      snc.dropTable(rr_table, ifExists = true)
+    }
     resetMemoryManagers
   }
